Restructure app to make it more readable/maintainable
